Move pontentials in new folder
